+++ src/include/gttcan.h
@@ -57,8 +57,8 @@ typedef struct gttcan_tag
     local_schedule_entry_t local_schedule[GTTCAN_MAX_LOCAL_SCHEDULE_LENGTH];
     // global_schedule_ptr_t global_schedule_ptr; //could keep a pointer in the future if we need to use global_schedule, i.e. fault tolerance or if global_schedule needs to be dynamically changed?
     uint16_t global_schedule_length;
-    uint8_t local_schedule_length;
-    uint8_t local_schedule_index;
+    uint16_t local_schedule_length;
+    uint16_t local_schedule_index;
     uint32_t slot_duration;
 
     // Callback functions
